requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==3.13.0
frozenlist==1.4.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
idna==3.6
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
twilio==8.2.2
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.3
